Moved cancel-drawing
