+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/cancel-drawing/cancel-drawing.tsx
